requirements.txt
beautifulsoup4==4.6.3
certifi==2018.4.16
chardet==3.0.4
googlemaps==3.0.2
html5lib==1.0.1
httplib2==0.12.0
idna==2.7
numpy==1.15.4
oauth2==1.9.0.post1
pandas==0.23.4
pymongo==3.7.2
python-dateutil==2.7.3
pytz==2018.7
PyYAML==3.13
requests==2.19.1
selenium==3.14.0
six==1.11.0
urllib3==1.23
webencodings==0.5.1
